+++ brocolli/converter/onnx_layers/avgpool_layer.py
@@ -33,7 +33,7 @@ class AvgPoolLayer(BaseLayer):
         elif isinstance(self._module, nn.AvgPool2d):
             pads = [0, 0, pad_h, pad_w, 0, 0, pad_h, pad_w]
 
-        params = [np.array(pads), np.array(0)]
+        params = [np.array(pads), np.array(0.0, dtype="float32")]
 
         pad_layer.generate_node(
             self._source_node.name + "_pad",
